README + Diagram
In this commit I've finished the draft of the GUI pages that we started previously, and I also made a neat diagram for visualization.

This README / Documentation does not contain the underlaying software processes, that handles the data or anything, this is just a High Level Design of the GUI.

Please review if I've made any errors, we should also discuss it because I feel like it could use some improvement. But let's see.

Also, I want more money.